move rasterize to helper module
